// admin/src/types/descubrePost.ts
export interface DescubrePostAdmin {
	id: string
	userId: string
	userName: string
	contactNumber: string
	moneda: string
	province: string
	precio: string
	date: string
	title: string
	description: string
	category: string
	imageUrls: string[]
	creditsSpent: number
	boostCredits: number
	status: string
	expiresAt: string
}